+++ src/dogs/core/dto/input.dto.ts
