+++ src/dogs/core/entities/dog.ts
