+++ src/dogs/core/interfaces/dogs.interfaces.ts
