+++ src/dogs/services/dogs.client.ts
